# Fix trace creation performance regression