--- a/R/call.R
+++ b/R/call.R
@@ -1057,22 +1057,10 @@ call_type <- function(x) {
 }
 
 call_zap_inline <- function(x) {
-  if (is_call(x)) {
-    if (is_call(x, "function")) {
-      if (!is_null(x[[2]])) {
-        x[[2]] <- as.pairlist(map(x[[2]], call_zap_inline))
-      }
-      x[[3]] <- call_zap_inline(x[[3]])
-    } else {
-      x[] <- map(x, call_zap_inline)
-    }
-    return(x)
-  }
+  .Call(ffi_call_zap_inline, x)
+}
 
-  if (is_symbol(x) || is_syntactic_literal(x)) {
-    x
-  } else {
-    name <- sprintf("<%s>", rlang_type_sum(x))
-    sym(name)
-  }
+# Called from C
+call_type_sum <- function(x) {
+  sym(sprintf("<%s>", rlang_type_sum(x)))
 }