--- a/R/trace.R
+++ b/R/trace.R
@@ -108,8 +108,12 @@ trace_back <- function(top = NULL, bottom = NULL) {
   calls <- map(calls, call_fix_car)
   calls <- map(calls, call_zap_inline)
 
-  context <- map2(calls, seq_along(calls), call_trace_context)
-  context <- inject(rbind(empty_trace_context(), !!!context))
+  context_data <- map2(calls, seq_along(calls), call_trace_context)
+
+  context <- empty_trace_context()
+  context$namespace <- do.call(base::c, map(context_data, `[[`, "namespace"))
+  context$scope <- do.call(base::c, map(context_data, `[[`, "scope"))
+  context <- new_data_frame(context)
 
   parents <- normalise_parents(parents)
 
@@ -240,7 +244,7 @@ call_trace_context <- function(call, fn) {
 }
 
 trace_context <- function(namespace = NA, scope = NA) {
-  data_frame(
+  list(
     namespace = namespace,
     scope = scope
   )