--- a/src/internal/decl/call-decl.h
+++ b/src/internal/decl/call-decl.h
@@ -1,2 +1,20 @@
 static
 bool is_callable(r_obj* x);
+
+static
+void call_zap_inline(r_obj* x);
+
+static
+void node_zap_inline(r_obj* x);
+
+static
+r_obj* call_zap_one(r_obj* x);
+
+static
+void call_zap_fn(r_obj* x);
+
+static
+r_obj* type_sum(r_obj* x);
+
+static
+r_obj* type_sum_call;