--- a/src/internal/exported.c
+++ b/src/internal/exported.c
@@ -849,25 +849,6 @@ r_obj* ffi_vec_poke_range(r_obj* x, r_obj* offset,
   return x;
 }
 
-static r_ssize validate_n(r_obj* n) {
-  if (n == r_null) {
-    return -1;
-  }
-
-  switch (r_typeof(n)) {
-  case R_TYPE_integer:
-  case R_TYPE_double:
-    if (r_length(n) == 1) {
-      break;
-    }
-    // fallthrough
-  default:
-    r_abort("`n` must be NULL or a scalar integer");
-  }
-
-  return r_arg_as_ssize(n, "n");
-}
-
 static int validate_finite(r_obj* finite) {
   switch (r_typeof(finite)) {
   case R_TYPE_null: