--- a/src/rlang/call.c
+++ b/src/rlang/call.c
@@ -30,6 +30,37 @@ r_obj* r_expr_protect(r_obj* x) {
   }
 }
 
+static inline
+bool is_node(r_obj* x) {
+  switch (r_typeof(x)) {
+  case R_TYPE_call:
+  case R_TYPE_pairlist:
+    return true;
+  default:
+    return false;
+  }
+}
+
+r_obj* r_call_clone(r_obj* x) {
+  if (!is_node(x)) {
+    r_abort("Input must be a call.");
+  }
+
+  x = KEEP(r_clone(x));
+
+  r_obj* rest = x;
+  while (rest != r_null) {
+    r_obj* head = r_node_car(rest);
+    if (is_node(head)) {
+      r_node_poke_car(rest, r_call_clone(head));
+    }
+    rest = r_node_cdr(rest);
+  }
+
+  FREE(1);
+  return x;
+}
+
 
 void r_init_library_call() {
   quote_prim = r_base_ns_get("quote");