Avoid data frames in a few places, and rewrite `call_zap_inline()` and `is_call()` in C.

This shrinks the benchmark below from 20ms to 1.1ms. This is still twice as slow as before (604µs), but after that point it becomes hard to optimise. The perf difference is due to the improved detection of namespaces and scopes stored in the backtrace.

cc @hadley.

```r
f <- function(n) {
  if (n) f(n - 1) else rlang::trace_back()
}
f(30)


bench::mark(f(30))[1:8]
# C `call_zap_inline()`
#>   expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#> 1 f(30)        1.06ms   1.14ms      830.    13.1KB     19.1   390     9

# C `is_call()`
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#> 1 f(30)        2.65ms 2.72ms      338.    15.2KB     27.6   147    12

# Faster df ops
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#> 1 f(30)        3.33ms 3.49ms      266.    13.1KB     18.1   118     8

# rlang 1.0.0
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#> 1 f(30)        19.2ms   20ms      50.1    30.2KB     142.     6    17

# rlang 0.4.12
#>   expression      min median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
#> 1 f(30)         583µs  604µs     1589.     558KB     99.7   669    42
```